refactor(fuzz): extract shared operation selection and paths preflight
